Seperate test in single files.
